# openapi.yaml
openapi: 3.0.0
info:
  title: Recipe API
  version: 1.0.0
paths:
  /recipes:
    get:
      summary: Get a list of all recipes
      responses:
        '200':
          description: A map of recipes
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/Recipe'
    post:
      summary: Create a new recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
      responses:
        '200':
          description: The created recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
  /recipes/{id}:
    get:
      summary: Get a single recipe by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A single recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '404':
          description: Recipe not found
    put:
      summary: Update an existing recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
      responses:
        '200':
          description: The updated recipe
        '404':
          description: Recipe not found
    delete:
      summary: Delete a recipe by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recipe deleted successfully
        '404':
          description: Recipe not found
components:
  schemas:
    Recipe:
      type: object
      properties:
        name:
          type: string
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
    Ingredient:
      type: object
      properties:
        name:
          type: string